Added redundancy fix to buedapy as well

+++ buedapy/recipes/default.rb
@@ -12,18 +12,21 @@ AWS::S3::Base.establish_connection!(
     :secret_access_key => node[:s3][:secret_access_key]
 )
 
-open("/tmp/buedapy.tar.gz", "w") do |file|
-  obj = AWS::S3::S3Object.find "buedapy.tar.gz", "bueda.deploy"
-  file.write obj.value
-end unless File.exists?("/tmp/buedapy.tar.gz")
+if not File.exists?(node[:buedapy][:virtualenv] 
+        + "/lib/python2.6/site-packages/language_model") then
+  open("/tmp/buedapy.tar.gz", "w") do |file|
+    obj = AWS::S3::S3Object.find "buedapy.tar.gz", "bueda.deploy"
+    file.write obj.value
+  end unless File.exists?("/tmp/buedapy.tar.gz")
 
-pip_package "buedapy" do
-  action :install_from_file
-  source "/tmp/buedapy.tar.gz"
-  virtualenv node[:buedapy][:virtualenv]
-  only_if do File.exists?("/tmp/buedapy.tar.gz") end
-end
+  pip_package "buedapy" do
+    action :install_from_file
+    source "/tmp/buedapy.tar.gz"
+    virtualenv node[:buedapy][:virtualenv]
+    only_if do File.exists?("/tmp/buedapy.tar.gz") end
+  end
 
-file "/tmp/buedapy.tar.gz" do 
-  action :delete
+  file "/tmp/buedapy.tar.gz" do 
+    action :delete
+  end
 end
